Extract image-adding logic to separate file
